// src/app/core/api/model/authzInstallResponseDto.ts
/**
 * Flui.cloud API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AuthzInstallResponseDto { 
    id: string;
    clusterId: string;
    clusterName: string;
    status: AuthzInstallResponseDto.StatusEnum;
    operationId?: string;
    errorMessage?: string;
    installedAt?: string;
    createdAt: string;
    updatedAt: string;
}
export namespace AuthzInstallResponseDto {
    export const StatusEnum = {
        Pending: 'PENDING',
        Installing: 'INSTALLING',
        Running: 'RUNNING',
        Failed: 'FAILED',
        Uninstalled: 'UNINSTALLED'
    } as const;
    export type StatusEnum = typeof StatusEnum[keyof typeof StatusEnum];
}